+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- Update README on sample apps
+
 ## Version 1.9.4 - 2022-08-01
 
 ### Fixed